requirements.txt
amqp==2.5.2
argon2-cffi==19.2.0
asgiref==3.2.3
autopep8==1.4.4
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.41.0
django-crispy-forms==1.8.1
django-environ==0.4.5
djangorestframework==3.11.0
djangorestframework-datatables==0.5.0
djangorestframework-gis==0.15
httpie==1.0.3
idna==2.8
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.5.2
python3-openid==3.1.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
